Tests/LibWeb/Text/input/css/scrollable-overflow-recalc-nested-propagation.html

<!DOCTYPE html>
<script src="../include.js"></script>
<div id="scroller" style="overflow: scroll; width: 100px; height: 100px;">
    <div id="wrapper" style="width: 50px; height: 50px;">
        <div id="child" style="width: 40px; height: 40px;"></div>
    </div>
</div>
<script>
    // A transformed box contributes to its scroll container through an intermediate
    // overflow: visible ancestor, so the recalculation has to propagate along the whole
    // containing block chain.
    test(() => {
        const scroller = document.getElementById("scroller");
        const child = document.getElementById("child");
        const initialHeight = scroller.scrollHeight;
        const countersBefore = internals.getStyleInvalidationCounters();
        child.style.transform = "translateY(400px)";
        const heightAfterChange = scroller.scrollHeight;
        const countersAfter = internals.getStyleInvalidationCounters();
        println(`initial scrollHeight: ${initialHeight}`);
        println(`scrollHeight after transform change: ${heightAfterChange}`);
        println(`relayouts performed: ${countersAfter.relayoutsPerformed - countersBefore.relayoutsPerformed}`);
    });
</script>
